@@ -1,29 +1,20 @@
 {
-  "name": "chat-app-client",
+  "name": "zhenye-web",
   "version": "0.1.0",
+  "homepage": "https://zhenyezhu.github.io",
   "private": true,
   "dependencies": {
-    "@apollo/client": "^3.5.10",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "bootstrap": "^5.1.3",
-    "classnames": "^2.3.1",
-    "graphql": "^16.3.0",
-    "jwt-decode": "^3.1.2",
-    "moment": "^2.29.2",
-    "node-sass": "^7.0.1",
-    "peerjs": "^1.3.2",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.2.1",
-    "react-dom": "^17.0.2",
-    "react-router": "^6.2.2",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
-    "subscriptions-transport-ws": "^0.11.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
@@ -46,5 +37,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "gh-pages": "^6.0.0"
   }
 }
